fix: json query

+++ pkg/describe/schedulers/discovery/old_resource_deleter.go
@@ -42,6 +42,14 @@ func (s *Scheduler) runDeleter() error {
 		switch task.Source.TaskType {
 		case es.DeleteTaskTypeResource:
 			for _, resource := range task.Source.DeletingResources {
+				job, err := s.db.GetDescribeConnectionJobByID(task.Source.DiscoveryJobID)
+				if err != nil {
+					s.logger.Error("failed to get describe connection job", zap.Error(err))
+					continue
+				}
+				if job.Status != api.DescribeResourceJobOldResourceDeletion {
+					continue
+				}
 				err = s.esClient.Delete(string(resource.Key), resource.Index)
 				if err != nil {
 					if strings.Contains(err.Error(), "[404 Not Found]") {
@@ -50,11 +58,6 @@ func (s *Scheduler) runDeleter() error {
 					}
 					s.logger.Error("failed to delete resource", zap.Error(err))
 					return err
-				}
-				job, err := s.db.GetDescribeConnectionJobByID(task.Source.DiscoveryJobID)
-				if err != nil {
-					s.logger.Error("failed to get describe connection job", zap.Error(err))
-					continue
 				}
 				err = s.db.UpdateDescribeConnectionJobStatus(job.ID, api.DescribeResourceJobSucceeded, job.FailureMessage, job.ErrorCode, job.DescribedResourceCount, job.DeletingCount)
 				if err != nil {
